[feat] posts from seller by user following list

+++ src/main/java/br/com/mercadolivre/bootcampw2/grupo11/socialmeli/services/UsersService.java
@@ -47,9 +47,9 @@ public class UsersService {
 
     public void follow(Integer customerId, Integer sellerId) {
         Customer tmpCustomer = customerRepository.findById(customerId)
-                .orElseThrow(() -> new ResourceNotFoundException("Customer", customerId.longValue()));
+                .orElseThrow(() -> new ResourceNotFoundException("Customer", customerId));
         Seller tmpSeller = sellerRepository.findById(sellerId)
-                .orElseThrow(() -> new ResourceNotFoundException("Seller", sellerId.longValue()));
+                .orElseThrow(() -> new ResourceNotFoundException("Seller", sellerId));
 
         tmpCustomer.addFollow(tmpSeller);
